+++ packages/studio/src/components/ForwardOAuth2Login/messages.tsx
@@ -1,0 +1,5 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  error: 'An error has occurred.',
+});