+++ packages/web-utils/src/types.ts
@@ -1,0 +1,24 @@
+/**
+ * An object that represents a named event target.
+ */
+export interface NamedEventTarget {
+  /**
+   * The name of the event target.
+   */
+  name?: string;
+}
+
+/**
+ * An event that refers to a named event target.
+ */
+export interface NamedEvent<T extends NamedEventTarget = NamedEventTarget, C = T> {
+  /**
+   * The event target.
+   */
+  target: T;
+
+  /**
+   * The current event target.
+   */
+  currentTarget?: C;
+}